Deck and joker constructor
Joker constructor takes a rank for uniqueness - rank doesn't really
change anything in terms of the "power" of a joker but helps with
ensuring jokers are unique in a deck.

Deck constructor takes an optional argument for the number of decks.
A deck includes two jokers, so n decks include 2n jokers.