Leave `FORMS` as is in `$` operator.
`$` takes a list of forms and makes a unary function which applies
them via `->>`.  Previous definition reversed the forms supplied
s.t. `$` was closer to the applicative operator in Haskell i.e. the
function (<$> f g) is f(g(x)).  But this version fits closer with the
lower operator (->>) being used AND allows easier lifting from
`->>` (which produces a value) to `$` (which produces a function).