Create a unit test system
Setup cantedraw/tests with a test-system record, as well as linking it
back to cantedraw so we can run test-system on cantedraw and run the
tests.

Using parachute, we can setup a suite of unit tests for each package.
For each package in src/ or lib/ we'll write a separate unit test
suite with tests for all the public exposing functionality.

I've setup a main suite in cantedraw/tests/main which will depend on
each of the unit test suites so we can run the tests all at once.

Currently we have one test for the --> macro.